requirements.txt
nats-py==2.6.0
redis==5.0.8
httpx==0.27.0
